Add and use a PluginLoader.
The pluginLoader loads and registers all available plugins in the plugin
dirs. This means it scans all of the plugin dirs before a render, but
only the used plugins are actually included.

The plugin loader enables plugins to be called with the dependency
injector, rather than directly. This makes Mako services available to
the plugins.